+++ Solutions/AtlassianJiraAudit/Data/Solution_AtlassianJiraAudit.json
@@ -31,16 +31,20 @@
     "Hunting Queries/JiraWorkflowAddedToProject.yaml"
   ],
   "Data Connectors": [
-    "Data Connectors/JiraNativePollerConnector/azuredeploy_Jira_native_poller_connector.json"
+    "Data Connectors/JiraAudit_API_FunctionApp.json"
   ],
   "Playbooks": [
     "Playbooks/Create-Jira-Issue/alert-trigger/azuredeploy.json",
     "Playbooks/Create-Jira-Issue/incident-trigger/azuredeploy.json",
     "Playbooks/Jira-CreateAndUpdateIssue/azuredeploy.json",
-    "Playbooks/Sync-AssignedUser/azuredeploy.json"
+    "Playbooks/Sync-AssignedUser/azuredeploy.json",
+    "Playbooks/Add-JiraLinkComment/azuredeploy.json",
+    "Playbooks/Sync-CommentsFunctionApp/azuredeploy.json",
+    "Playbooks/Sync-Incidents/azuredeploy.json",
+    "Playbooks/Sync-Status/azuredeploy.json"
   ],
   "BasePath": "C:\\GitHub\\azure-Sentinel\\Solutions\\AtlassianJiraAudit",
-  "Version": "2.0.0",
+  "Version": "2.0.3",
   "Metadata": "SolutionMetadata.json",
   "TemplateSpec": true,
   "Is1PConnector": false
