+++ Solutions/AtlassianJiraAudit/Hunting Queries/JiraUserIPs.yaml
@@ -15,7 +15,7 @@ query: |
   JiraAudit
   | where TimeGenerated > ago(24h)
   | where isnotempty(SrcIpAddr)
-  | where isnotempty(USerName)
+  | where isnotempty(UserName)
   | summarize ip_list = makeset(SrcIpAddr) by UserName
   | extend AccountCustomEntity = UserName
 entityMappings:
